♻️ refactor:  파일명 postMessage로 변경
